api/requirements.txt
numpy==1.26.0
Flask==3.0.0
marshmallow==3.20.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
flask-cors==4.0.0
python-dotenv==1.0.0
